requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==4.2.16
django-bootstrap-v5==1.0.11
django-tinymce==4.1.0
gunicorn==20.1.0
psycopg2==2.9.10
sqlparse==0.5.1